refactor: VoteResponse record로 변경 및 Converter에서 변환하도록 변경
