+++ src/main/java/com/gamemoonchul/infrastructure/web/dto/response/VoteRatioResponse.java
@@ -1,16 +1,11 @@
 package com.gamemoonchul.infrastructure.web.dto.response;
 
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
-@Getter
 @Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class VoteRatioResponse {
-    private MatchUserResponse matchUserResponse;
-    private Double voteRatio;
+public record VoteRatioResponse(
+    MatchUserResponse matchUserResponse,
+    Double voteRatio
+) {
+
 }
